renamed gallery page to about-us and moved image rotation logic to AboutUsGallery component
